homework 1/8

+++ resources/grower/routes.py
@@ -1,5 +1,6 @@
 from flask import request
 from flask.views import MethodView
+from flask_jwt_extended import get_jwt_identity, jwt_required
 from flask_smorest import abort
 
 from models.growerModel import GrowerModel
@@ -17,18 +18,21 @@ class Grower(MethodView):
         else:
             abort(400, message="Grower not found")
 
+    @jwt_required()
     @bp.arguments(GrowerSchema)
     def put(self, grower_data, grower_name):
-        grower=GrowerModel.query.get(grower_name)
-        if grower:
+        grower=GrowerModel.query.get(get_jwt_identity())
+        if grower and grower.grower_name==grower_name:
             grower.from_entry(grower_data)
             grower.commit()
+            print (grower.grower_name)
             return{'message':f'{grower.grower_name} updated'},202
         abort(400, message='Invalid Grower Name')
 
+    @jwt_required()
     def delete(self, grower_name):
-        grower=GrowerModel.query.get(grower_name)
-        if grower:
+        grower=GrowerModel.query.get(get_jwt_identity())
+        if grower and grower.grower_name==grower_name:
             grower.delete()
             return{'message':f'{grower.grower_name} deleted'},202
         abort(400, message='Invalid Grower Name')
@@ -46,6 +50,6 @@ class GrowerList(MethodView):
             grower=GrowerModel()
             grower.from_entry(grower_data)
             grower.commit()
-            return{'message':f'{grower.grower_name} Created'},201
+            return{'message' : f'{grower.grower_name} created'},201
         except:
             abort(400, message='Grower Name Already In System')
